requirements.txt
alembic==1.13.3
bleach==6.1.0
blinker==1.8.2
Brotli==1.1.0
cffi==1.17.1
click==8.1.7
cssselect2==0.7.0
flask==3.0.3
Flask-Login==0.6.3
flask-mail==0.10.0
Flask-Migrate==4.0.7
Flask-Redmail==0.3.0
Flask-Script==2.0.6
Flask-SeaSurf==2.0.0
flask-sqlalchemy==3.1.1
flask-talisman==1.1.0
flask-weasyprint==1.0.0
flask-wtf==1.2.1
fonttools==4.54.1
greenlet==3.1.1
gunicorn==23.0.0
html5lib==1.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
pillow==10.4.0
pycparser==2.22
pydyf==0.6.0
PyJWT==2.9.0
PyMySQL==1.1.1
pyphen==0.16.0
python-dotenv==1.0.1
redmail==0.6.0
six==1.16.0
SQLAlchemy==2.0.35
tinycss2==1.3.0
typing-extensions==4.12.2
weasyprint==59.0
webencodings==0.5.1
werkzeug==3.0.4
wtforms==3.1.2
WTForms-SQLAlchemy==0.4.1
zipp==3.20.2
zopfli==0.2.3
